fixup! SystemUI: Add support for timeout-reboot

+++ packages/SystemUI/src/com/android/systemui/keyguard/KeyguardViewMediator.java
@@ -1172,7 +1172,7 @@ public class KeyguardViewMediator extends SystemUI implements Dumpable,
         final ContentResolver cr = mContext.getContentResolver();
 
         // From Global Settings
-        final long timeout = LineageSettings.Global.getInt(cr,
+        final long timeout = LineageSettings.Global.getLong(cr,
                 LineageSettings.Global.DEVICE_REBOOT_TIMEOUT,
                 KEYGUARD_REBOOT_AFTER_DELAY_DEFAULT);
